Fix lazy-load Swiper via client component wrapper
Move next/dynamic call from Server Component (where it was ignored)
into a new LazyProductCarousel "use client" wrapper. This properly
code-splits Swiper (~50KiB) into a lazy chunk that loads after initial
paint, unblocking mobile hero LCP. SSR still renders carousel HTML
with product images for desktop LCP.
